pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.27",
    "celery>=5.6.2",
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-storages>=1.14.6",
    "django-unfold>=0.76.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=23.0.0",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "whitenoise>=6.11.0",
]
